requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
blinker==1.6.2
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.3.1
frozenlist==1.3.3
identify==2.5.23
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
pre-commit==3.2.2
psycopg2-binary==2.9.6
PyJWT==2.6.0
PySocks==1.7.1
pytest==7.3.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.29.0
six==1.16.0
tomli==2.0.1
twilio==8.1.0
urllib3==1.26.15
virtualenv==20.23.0
Werkzeug==2.3.1
yarl==1.9.2
